fix: delete tour

+++ src/Explorer.API/Controllers/Author/Administration/TourController.cs
@@ -40,7 +40,7 @@ namespace Explorer.API.Controllers.Author.Administration
         }
 
         [HttpDelete("{id:int}")]
-        public ActionResult Delete(int id)
+        public ActionResult Delete(long id)
         {
             var result = _tourService.Delete(id);
             return CreateResponse(result);
