+++ src/Modules/Tours/Explorer.Tours.API/Public/Administration/ITourService.cs
@@ -14,6 +14,6 @@ namespace Explorer.Tours.API.Public.Administration
         Result<PagedResult<TourDto>> GetPaged(int page,int pageSize);
         Result<TourDto> Create(TourDto tour);
         Result<TourDto> Update(TourDto tour);
-        Result Delete(int id);
+        Result Delete(long id);
     }
 }
